@@ -21,21 +21,21 @@
     "node-sass": "4.5.3",
     "sass-loader": "6.0.6",
     "style-loader": "0.19.0",
-    "ts-loader": "2.3.7",
+    "ts-loader": "3.0.3",
     "ts-node": "3.3.0",
     "ts-node-dev": "1.0.0-pre.4",
     "tslint": "5.7.0",
     "typescript": "2.5.3",
     "url-loader": "0.6.2",
-    "webpack": "3.7.1",
+    "webpack": "3.8.1",
     "webpack-closure-compiler": "2.1.5",
-    "webpack-dev-server": "2.9.1"
+    "webpack-dev-server": "2.9.2"
   },
   "dependencies": {
     "@types/config": "0.0.33",
     "@types/material-ui": "0.18.3",
-    "@types/node": "8.0.34",
-    "@types/react": "16.0.10",
+    "@types/node": "8.0.45",
+    "@types/react": "16.0.14",
     "@types/react-dom": "16.0.1",
     "@types/react-redux": "5.0.10",
     "@types/react-responsive": "^1.1.8",
@@ -46,8 +46,8 @@
     "@types/socket.io": "^1.4.31",
     "@types/socket.io-client": "^1.4.31",
     "@types/sqlite3": "3.1.1",
-    "@types/winston": "^2.3.6",
-    "config": "1.26.2",
+    "@types/winston": "2.3.6",
+    "config": "1.27.0",
     "core-js": "2.5.1",
     "immutable": "4.0.0-rc.7",
     "material-ui": "0.19.4",
@@ -63,11 +63,12 @@
     "react-tap-event-plugin": "3.0.2",
     "redux": "3.7.2",
     "redux-observable": "0.16.0",
-    "rxjs": "5.4.3",
+    "rxfiddle": "^1.0.4",
+    "rxjs": "5.5.0",
     "socket.io": "^2.0.3",
     "socket.io-client": "^2.0.3",
     "sqlite3": "3.1.13",
-    "typeface-roboto": "0.0.35",
-    "winston": "^2.4.0"
+    "typeface-roboto": "0.0.41",
+    "winston": "2.4.0"
   }
 }
