@@ -1,20 +1,21 @@
 import "core-js";
 
-import { Observable } from "rxjs/Rx";
+import * as Rx from "rxjs/Rx";
 
 import { onBlogApiRequest } from "./blog/http-request-handler.func";
 import { onFileRequest } from "./file-system/http-request-handler.func";
 import { loginfo } from "./logger/logger";
-import { openSocketServer, SocketRequestHandler } from "./sockets/server";
+import { openServer } from "./server/open-server.func";
+import { SocketHttpRequestHandler } from "./sockets/server";
 
 const port = 3000;
 
-const handler: SocketRequestHandler = (request$, sockets$) => Observable.concat(
-	onBlogApiRequest(request$),
-	onFileRequest(request$)
+const handler: SocketHttpRequestHandler = (request) => Rx.Observable.concat(
+	onBlogApiRequest(request),
+	onFileRequest(request)
 );
 
-openSocketServer(port, handler).subscribe(
+openServer(port, handler).subscribe(
 	server => { },
 	(err: Error) => console.error(err),
 	() => loginfo("...server closed.")
