@@ -1,19 +1,17 @@
 import { logerr, loginfo } from "../logger/logger";
 import { HttpRequest } from "./http-request.type";
 import { HttpRequestHandler } from "./http-request-handler.type";
-import { createServer, IncomingMessage, Server } from "http";
+import { createServer, IncomingMessage, Server, ServerResponse } from "http";
 import { List, Map } from "immutable";
-import { Observable } from "rxjs/Rx";
+import { Observable, ReplaySubject, Subscription, Subject } from "rxjs/Rx";
 import { Observer } from "rxjs/Observer";
 
 import { HttpResponse } from "./http-response.type";
-import { Subscription } from "rxjs/Subscription";
 
 export function openServer(port: number, handler: HttpRequestHandler): Observable<Server> {
-	return Observable.create((observer: Observer<Server>) => {
-		const server = createServer((request, response) => {
-			loginfo("Processing request", { request_url: request.url, request_method: request.method });
-
+	const requestResponse = new Subject<{ request: IncomingMessage, response: ServerResponse }>();
+	const pipeline = requestResponse
+		.mergeMap(({ request, response }) => {
 			const toArrayIfNot = <T>(v: T | T[]): T[] => Array.isArray(v) ? v : [v];
 			const headers = Map(Object.keys(request.headers)
 				.map(key => ({ [key]: List(toArrayIfNot(request.headers[key])) }))
@@ -23,7 +21,7 @@ export function openServer(port: number, handler: HttpRequestHandler): Observabl
 			const url = request.url || "/";
 
 			// Only connect to the body stream once but not until actually reading (so GET should not ever need to subscribe)
-			const bodyInner = (Observable.create((observer: Observer<Buffer>) => {
+			const body = (Observable.create((observer: Observer<Buffer>) => {
 				const onData = (chunk: Buffer | string) => observer.next(Buffer.isBuffer(chunk) ? chunk : new Buffer(chunk));
 				const onComplete = () => observer.complete();
 				const onError = (err: Error) => observer.error(err);
@@ -33,18 +31,14 @@ export function openServer(port: number, handler: HttpRequestHandler): Observabl
 					.on("error", onError);
 				return () => { };
 			}) as Observable<Buffer>).publishReplay();
+			const bodySub = body.connect();
 
-			let subscription: Subscription | null = null;
-			const body: Observable<Buffer> = Observable.create((observer: Observer<Buffer>) => {
-				const bodySubscription = bodyInner.subscribe(observer);
-				if (subscription == null) {
-					subscription = bodyInner.connect();
-				}
-				return () => bodySubscription.unsubscribe();
-			});
-			const responseSub = handler(Observable.of({ body, method, url, headers }))
+			const breaker = new Subject<{}>();
+			request.addListener("close", () => breaker.next({}));
+			return handler({ body, method, url, headers })
 				.defaultIfEmpty({ status: 404, body: JSON.stringify({ error: "NotFound", code: 404 }) })
 				.take(1)
+				.takeUntil(breaker)
 				.catch(error => {
 					logerr(`Unexpected error occured: ${error.message}`, { error, request_method: request.method, request_url: request.url });
 					return Observable.of({
@@ -57,8 +51,7 @@ export function openServer(port: number, handler: HttpRequestHandler): Observabl
 						headers: Map<string, List<string>>()
 					});
 				})
-				.finally(() => subscription && subscription.unsubscribe())
-				.subscribe(res => {
+				.do(res => {
 					response.statusCode = res.status || 404;
 					if (res.headers) {
 						res.headers.forEach((values, key) => {
@@ -68,11 +61,21 @@ export function openServer(port: number, handler: HttpRequestHandler): Observabl
 					if (res.body) {
 						response.write(res.body);
 					}
-					response.end();
 
 					loginfo("Response", { response_statuscode: response.statusCode, request_method: request.method, request_url: request.url });
+				})
+				.finally(() => {
+					response.end();
+					bodySub.unsubscribe();
 				});
-			request.addListener("close", () => responseSub.unsubscribe());
+		})
+		.subscribe();
+
+	return Observable.create((observer: Observer<Server>) => {
+		const server = createServer((request, response) => {
+			loginfo("Processing request", { request_url: request.url, request_method: request.method });
+
+			requestResponse.next({ request, response });
 		});
 
 		server.listen(port, (error: Error) => {
@@ -86,6 +89,7 @@ export function openServer(port: number, handler: HttpRequestHandler): Observabl
 
 		observer.next(server);
 		return () => {
+			pipeline.unsubscribe();
 			server.close();
 		};
 	});
